Fix regular failure in offset, limit happy case because of Natural and Positive generation

+++ core/src/test/scala/quasar/fs/NumericGen.scala
@@ -5,6 +5,6 @@ import quasar.Predef._
 
 object NumericGen {
   // TODO: Replace use of get with `flatten` or `collect` once we upgrade scalacheck
-  implicit val arbPositive = Arbitrary.apply(Arbitrary.arbitrary[Long].filter(_ > 0).map(long => Positive(long).get))
-  implicit val arbNatural = Arbitrary.apply(Arbitrary.arbitrary[Long].filter(_ >= 0).map(long => Natural(long).get))
+  implicit val arbPositive = Arbitrary.apply(Arbitrary.arbitrary[Short].filter(_ > 0).map(long => Positive(long).get))
+  implicit val arbNatural = Arbitrary.apply(Arbitrary.arbitrary[Short].filter(_ >= 0).map(long => Natural(long).get))
 }
